+++ logger/hourai-logger/src/message_logging.rs
@@ -21,7 +21,7 @@ fn message_base_embed(message: &impl MessageLike) -> Result<EmbedBuilder> {
         )
         .title(format!("ID: {}", message.id()))
         .url(message.message_link())
-        .timestamp(Timestamp::from_secs(Utc::now().timestamp() as u64).unwrap()))
+        .timestamp(Timestamp::from_secs(Utc::now().timestamp()).unwrap()))
 }
 
 pub(crate) fn message_to_embed(message: &impl MessageLike) -> Result<EmbedBuilder> {
